move float to core.function
