+++ sympy/solvers/solvers.py
@@ -12,12 +12,12 @@
 """
 
 from sympy.core.compatibility import iterable, is_sequence
-from sympy.core.containers import Dict
 from sympy.core.sympify import sympify
 from sympy.core import C, S, Mul, Add, Pow, Symbol, Wild, Equality, Dummy, Basic, Expr
 from sympy.core.function import (expand_mul, expand_multinomial, expand_log,
-        Derivative, Function, AppliedUndef, UndefinedFunction, count_ops)
-from sympy.core.numbers import ilcm, Float, Rational
+        Derivative, Function, AppliedUndef, UndefinedFunction, count_ops,
+	float)
+from sympy.core.numbers import ilcm, Float
 from sympy.core.relational import Relational
 from sympy.logic.boolalg import And, Or
 
@@ -46,7 +46,6 @@ from warnings import warn
 from textwrap import fill, dedent
 from types import GeneratorType
 from collections import defaultdict
-float_py = float
 
 # if you use
 # _filldedent('''
@@ -304,70 +303,6 @@ def check_assumptions(expr, **assumptions):
             return False
         result = None # Can't conclude, unless an other test fails.
     return result
-
-def float(expr, denom_of_1=False, exponent=False):
-    """Make all Rationals in expr Floats except if they are exponents
-    (unless the exponents flag is set to True). If denom_of_1 is False
-    then rational coefficients with a numerator of 1 will not be changed.
-
-    Examples:
-
-    >>> from sympy.solvers.solvers import float
-    >>> from sympy.abc import x
-    >>> from sympy import cos, pi, S
-    >>> float(x**4 + x/2 + cos(pi/3) + 1)
-    x**4 + x/2 + 1.5
-    >>> float(x**4 + x/2 + cos(pi/3) + 1, denom_of_1=True, exponent=True)
-    0.5*x + x**4.0 + 1.5
-
-    """
-
-    if iterable(expr, exclude=basestring):
-        if isinstance(expr, (dict, Dict)):
-            return type(expr)([(k, float(v, denom_of_1, exponent)) for k, v in expr.iteritems()])
-        return type(expr)([float(a, denom_of_1, exponent) for a in expr])
-    elif not isinstance(expr, Expr):
-        return float_py(expr)
-    elif expr.is_Float:
-        return expr
-    elif expr.is_Integer:
-        return Float(float_py(expr))
-    elif expr.is_Rational:
-        return Float(expr)
-
-    denoms = []
-    if not denom_of_1:
-        denoms = [(r, Dummy()) for r in expr.atoms(Rational) if r.p == 1 and r.q != 1]
-
-    if exponent is False:
-        pows = [p for p in expr.atoms(Pow) if p.exp.is_Rational and p.exp.q != 1]
-        pows.sort(key=count_ops)
-        pows.reverse()
-        rats = {}
-        for p in pows:
-            if p.exp not in rats:
-                e = Dummy()
-                rats[p.exp] = e
-        reps = [(p, Pow(p.base, rats[p.exp], evaluate=False)) for p in pows]
-        rv = expr.subs(reps).subs(denoms).n()
-        rv = rv.subs([(v, k) for k, v in rats.iteritems()]).subs([(n, o) for o, n in denoms])
-    else:
-        expr = expr.subs(denoms).n().subs([(n, o) for o, n in denoms])
-        if exponent is True:
-            pows = [p for p in expr.atoms(Pow) if p.exp.is_Integer]
-            pows.sort(key=count_ops)
-            pows.reverse()
-            ints = {}
-            for p in pows:
-                if p.exp not in ints:
-                    ints[p.exp] = Float(float_py(p.exp))
-            reps = [(p, p.base**ints[p.exp]) for p in pows]
-            rv = expr.subs(reps)
-
-    funcs = [f for f in rv.atoms(Function)]
-    funcs.sort(key=count_ops)
-    funcs.reverse()
-    return rv.subs([(f, f.func(*[float(a, denom_of_1, exponent) for a in f.args])) for f in funcs])
 
 def solve(f, *symbols, **flags):
     """
